Dir watch (#45)
* Setup chokidar

* Update config

* Notes and pathing

* notes

* On unlink function

* Add file

* Specify node version

* Diff directories to see if file is removed or updated

* Remove on change

* Update so watcher runs on setup from root

* 1.17.0

* Update build to use node 20

* RM yml ndoe

* Remove electron-dl

* Testing progress loading models

* RM console

* Fix model loading progress bar

* Fix some hashing to lowercase

* Remove consoles

* Update default root path in case it doesnt exist

* 1.17.0-beta.1

* Dont toast remove resource

* Typecheck fix

* Switch to delay

* 1.17.0-beta.2

* Fixing search

* Update image extension

* Replace resourceRemove with deleteFile to prevent it removing files, probably need better names

* 1.17.0-beta.3

* Change delete file function for watcher too

* 1.17.0-beta.4

* Update folder-watcher

* 1.17.0-beta.5

* Update mac build for m1

* beta 6

* Prevent removing files from store on folder change

* Offload hashing to worker thread

* Update with thread worker

* Update startup files and folder watching startup

* Filter file name better

+++ src/main/commands/resources-remove.ts
@@ -4,9 +4,10 @@ import { getResourcePath } from '../store/paths';
 import { deleteFile, searchFile } from '../store/files';
 import { updateActivity } from '../store/activities';
 import { filterResourcesList } from './filter-reources-list';
+import { getUrlExtension } from '../utils/get-url-extension';
 
 export function resourcesRemove(hash: string) {
-  const resource = searchFile(hash);
+  const resource = searchFile(hash.toLowerCase());
   const defaultResourcePath = getResourcePath(resource.type);
   const timestamp = new Date().toISOString();
   const resourcePath =
@@ -14,12 +15,17 @@ export function resourcesRemove(hash: string) {
 
   try {
     // Remove model from disk
-    fs.unlinkSync(resourcePath);
+    if (fs.existsSync(resourcePath)) fs.unlinkSync(resourcePath);
 
-    // Remove thumbnail from disk
-    const previewPath =
-      resourcePath.split('.').slice(0, -1).join('.') + '.preview.png';
-    if (fs.existsSync(previewPath)) fs.unlinkSync(previewPath);
+    if (resource.previewImageUrl) {
+      // Remove thumbnail from disk
+      const extension = getUrlExtension(resource.previewImageUrl);
+      const previewPath =
+        resourcePath.split('.').slice(0, -1).join('.') +
+        '.preview.' +
+        extension;
+      if (fs.existsSync(previewPath)) fs.unlinkSync(previewPath);
+    }
 
     // Remove json data from disk
     const jsonPath = resourcePath.split('.').slice(0, -1).join('.') + '.json';
